+++ docs/CHANGELOG.md
@@ -2,3 +2,6 @@
 + Remove the mechanism to manage non-standardized prj file
 + DBFReader uses VarChar instead of Char for String values
 + Change move some log info to debug mode
++ Use the next generation overlay
++ Add grid size precision to snap intersection, difference and symdifference operators
++ Improve ST_Union to use CoverageUnion when the input geometry contains only polygons